+++ general/mytransforms.h
@@ -31,14 +31,10 @@ class MyTransforms
   float *dataTemp;
   float *dataTime, *dataFFT;
   float *autocorrTime, *autocorrFFT;
-  //float *storeFFT; //stores a copy of the fft done during the autocorr calculation
   float *hanningCoeff, hanningScalar;
   float *harmonicsAmpLeft, *harmonicsPhaseLeft;
   float *harmonicsAmpCenter, *harmonicsPhaseCenter;
   float *harmonicsAmpRight, *harmonicsPhaseRight;
-  //float *equalLoudnessTable;
-  //float *storeFFTAmp1;
-  //float *storeFFTAmp2;
   bool beenInit;
   double freqPerBin;
   //float threshold;
@@ -58,22 +54,19 @@ class MyTransforms
   double autoLogCorr(float *input, float *output);
   double asdf(float *input, float *output);
   double nsdf(float *input, float *output);
-  //void calculateAnalysisData(float *input, AnalysisData &analysisDatam, Channel *ch, float threshold);
-  static int findNSDFMaxima(float *input, int len, std::vector<int> &maxPositions);
-  static int findNSDFsubMaximum(float *input, int len, float threshold);
-  void calculateAnalysisData(/*float *input, */int chunk, Channel *ch/*, float threshold*/);
-  //void buildEqualLoudnessTable(double dB=50.0);
-  //void doEqualLoudness();
+  static size_t findNSDFMaxima(float *input, size_t len, std::vector<size_t> &maxPositions);
+  static size_t findNSDFsubMaximum(float *input, size_t len, float threshold);
+  void calculateAnalysisData(int chunk, Channel *ch/*, float threshold*/);
+
   void doChannelDataFFT(Channel *ch, float *curInput, int chunk);
   void calcHarmonicAmpPhase(float *harmonicAmp, float *harmonicPhase, int binsPerHarmonic);
   void doHarmonicAnalysis(float *input, AnalysisData &analysisData, double period);
-  //void chooseCorrelationIndex(AnalysisData &analysisData, float threshold, float periodOctaveEstimate);
-  //void applyHighPassFilter(float *input, float *output);
+
   float get_fine_clarity_measure(double period);
   double get_max_note_change(float *input, double period);
   void applyHanningWindow(float *d);
-  static double calcFreqCentroid(float *buffer, int len);
-  static double calcFreqCentroidFromLogMagnitudes(float *buffer, int len);
+  static double calcFreqCentroid(float *buffer, size_t len);
+  static double calcFreqCentroidFromLogMagnitudes(float *buffer, size_t len);
 };
   
 #endif
